+++ CHANGES.rst
@@ -24,7 +24,8 @@ Changelog
 - Make fancybox organization gallery scrollable : #20754
   [laulaz]
 
-- Add 'hover' class when organization image changes (with delay) : #20754
+- Add 'hover' and 'hover-delay' (with delay) class when organization image
+  changes : #20754
   [laulaz]